fix for incorrect total value/supply

+++ Sources/daemon/database.swift
@@ -309,7 +309,7 @@ SUM(d5000) as d5000 FROM stats;
                 rpcstats.depletion = summary_stats.results[0]["depletion"]!.asDouble() ?? 0
                 rpcstats.height = summary_stats.results[0]["blocks"]!.asInt() ?? 0
                 rpcstats.tokens = summary_stats.results[0]["tokens"]!.asInt() ?? 0
-                rpcstats.value = Double(summary_stats.results[0]["value"]!.asInt() ?? 0)
+                rpcstats.value = summary_stats.results[0]["value"]!.asDouble() ?? 0
                 rpcstats.rate = summary_stats.results[0]["rate"]!.asDouble() ?? 0
                 rpcstats.transactions = summary_stats.results[0]["transactions"]!.asInt() ?? 0
                 rpcstats.addresses = summary_stats.results[0]["addresses"]!.asInt() ?? 0
